Styles refactored
