+++ src/components/Navbar/NavbarDropdown.module.scss
@@ -1,0 +1,19 @@
+.text {
+  font-family: Rubik, sans-serif;
+  font-style: normal;
+  font-weight: normal;
+  line-height: 20px;
+  font-size: 14px;
+  color: #15181f;
+}
+
+.list {
+  padding: 8px;
+  a {
+    text-decoration: none;
+  }
+}
+
+.item {
+  padding: 6px 8px;
+}
