--- src/components/Navbar/NavbarHelpDropdown.module.scss
@@ -1,31 +1,0 @@
-@import '~@santiment-network/ui/variables.scss';
-
-.text {
-  font-family: Rubik, sans-serif;
-  font-style: normal;
-  font-weight: normal;
-  line-height: 20px;
-  font-size: 14px;
-  color: #15181f;
-}
-
-.wrapper {
-  background-color: #fff;
-  border: 1px solid #e0e4ee;
-  box-shadow: 0px 1px 2px rgba(21, 24, 31, 0.02),
-    0px 2px 10px rgba(21, 24, 31, 0.04), 0px 4px 24px rgba(21, 24, 31, 0.06);
-  border-radius: $border-radius;
-  width: 260px;
-  box-sizing: border-box;
-  & * {
-    box-sizing: border-box;
-  }
-}
-
-.list {
-  padding: 8px;
-}
-
-.item {
-  padding: 6px 8px;
-}
