+++ src/components/Navbar/NavbarProfileDropdown.module.scss
@@ -1,44 +1,3 @@
-@import '~@santiment-network/ui/variables.scss';
-
-// @NOTE(vanguard): change this
-ul {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-h3,
-h4 {
-  margin: 0;
-}
-
-@mixin menu-side-padding {
-  padding-left: 20px;
-  padding-right: 16px;
-}
-
-.text {
-  font-family: Rubik, sans-serif;
-  font-style: normal;
-  font-weight: normal;
-  line-height: 20px;
-  font-size: 14px;
-  color: #15181f;
-}
-
-.wrapper {
-  background-color: #fff;
-  border: 1px solid #e0e4ee;
-  box-shadow: 0px 1px 2px rgba(21, 24, 31, 0.02),
-    0px 2px 10px rgba(21, 24, 31, 0.04), 0px 4px 24px rgba(21, 24, 31, 0.06);
-  border-radius: $border-radius;
-  width: 260px;
-  box-sizing: border-box;
-  & * {
-    box-sizing: border-box;
-  }
-}
-
 .onlineIndicator {
   border-radius: 50%;
   width: 6px;
@@ -58,8 +17,7 @@ h4 {
 }
 
 .profile {
-  padding: 15px 0 12px;
-  @include menu-side-padding;
+  padding: 15px 20px 12px 16px;
 
   &__pic {
     width: 32px;
@@ -77,7 +35,12 @@ h4 {
     }
   }
 
+  &__name {
+    margin: 0;
+  }
+
   &__status {
+    margin: 0;
     line-height: 20px;
     font-size: 12px;
     color: #7a859e;
@@ -101,14 +64,6 @@ h4 {
 
   &__amount {
     color: #26292f;
-  }
-}
-
-.menuList {
-  padding: 8px;
-
-  &__item {
-    padding: 6px 8px;
   }
 }
